Recreated demo
After recreated PyLights to be less of an application and more of a tool
for developers, I also needed to redesign my demo. The demo is now
easier to pair with your philips hue bridge, allows you to add lights
and music while running, and is based on PyLights instead of repeating
code inside of PyLights.